Oraculo/pyproject.toml
[project]
name = "oraculo"
version = "0.1.0"
description = "API module for deploying MoE-based IDS"
authors = [
    {name = "Damitz"},
    {name = "Bronson"},
    {name = "Daniel Cuadros"},
    {name = "Fabricio Asfora",email = "asfora.assuncao@ime.eb.br"},
    {name = "Roosevelth Soares", email = "roosevelth.soares@ime.eb.br"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "absl-py (==2.1.0)",
    "annotated-types (==0.7.0)",
    "astunparse (==1.6.3)",
    "blinker (==1.6.2)",
    "certifi (==2023.7.22)",
    "charset-normalizer (==3.2.0)",
    "click (==8.1.7)",
    "flask (==3.0.0)",
    "flatbuffers (==25.2.10)",
    "gast (==0.6.0)",
    "google-pasta (==0.2.0)",
    "grpcio (==1.70.0)",
    "h5py (==3.13.0)",
    "idna (==3.4)",
    "itsdangerous (==2.1.2)",
    "jinja2 (==3.1.2)",
    "joblib (==1.3.2)",
    "keras (==3.8.0)",
    "libclang (==18.1.1)",
    "markdown (==3.7)",
    "markdown-it-py (==3.0.0)",
    "markupsafe (==2.1.3)",
    "mdurl (==0.1.2)",
    "ml-dtypes (==0.4.1)",
    "namex (==0.0.8)",
    "numpy (==1.26.0)",
    "opt-einsum (==3.4.0)",
    "optree (==0.14.0)",
    "packaging (==24.2)",
    "pandas (==2.1.1)",
    "pika (==1.3.2)",
    "protobuf (==5.29.3)",
    "pyarrow (==20.0.0)",
    "pydantic (==2.10.6)",
    "pydantic-core (==2.27.2)",
    "pygments (==2.19.1)",
    "python-dateutil (==2.8.2)",
    "python-dotenv (==1.0.0)",
    "pytz (==2023.3.post1)",
    "requests (==2.31.0)",
    "rich (==13.9.4)",
    "scikit-learn (==1.3.2)",
    "scipy (==1.11.2)",
    "six (==1.16.0)",
    "tensorboard (==2.18.0)",
    "tensorboard-data-server (==0.7.2)",
    "tensorflow (==2.18.0)",
    "tensorflow-io-gcs-filesystem (==0.37.1)",
    "termcolor (==2.5.0)",
    "threadpoolctl (==3.2.0)",
    "tqdm (==4.67.1)",
    "typing-extensions (==4.12.2)",
    "tzdata (==2023.3)",
    "urllib3 (==2.0.5)",
    "werkzeug (==3.0.0)",
    "wrapt (==1.17.2)",
    "mlflow (>=3.2.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
